+++ src/readers/dao.py
@@ -1,6 +1,10 @@
 from sqlalchemy import insert, select
 
 from src.database import async_session_maker
+from src.exceptions.user_exceptions import (
+    TheReaderAlreadyExistsException,
+    TheReaderWasNotFoundException,
+)
 from src.readers.models import Reader
 
 
@@ -18,32 +22,46 @@ class ReaderDAO:
     @classmethod
     async def add_reader(cls, **data):
         async with async_session_maker() as session:
+            active_reader_query = select(cls.model).filter(
+                cls.model.email == data["email"]
+            )
+            active_reader = await session.execute(active_reader_query)
+            active_reader = active_reader.scalar_one_or_none()
+            if active_reader:
+                raise TheReaderAlreadyExistsException
             query = insert(cls.model).values(**data)
             await session.execute(query)
             await session.commit()
 
     @classmethod
-    async def delete_reader(cls, reader_id):
+    async def delete_reader(cls, reader_id: int):
         async with async_session_maker() as session:
             query = select(cls.model).filter(cls.model.id == reader_id)
             row = await session.execute(query)
-            row = row.scalar_one()
+            row = row.scalar_one_or_none()
+            if row is None:
+                raise TheReaderWasNotFoundException
             await session.delete(row)
             await session.commit()
 
     @classmethod
-    async def update_reader(cls, reader_id, **data):
+    async def update_reader(cls, reader_id: int, **data):
         async with async_session_maker() as session:
             query = select(cls.model).filter(cls.model.id == reader_id)
             row = await session.execute(query)
-            row = row.scalar_one()
+            row = row.scalar_one_or_none()
+            if row is None:
+                raise TheReaderWasNotFoundException
             for key, value in data.items():
                 setattr(row, key, value)
             await session.commit()
 
     @classmethod
-    async def find_by_id(cls, reader_id: int):
+    async def show_reader(cls, reader_id: int):
         async with async_session_maker() as session:
             query = select(cls.model).filter_by(id=reader_id)
-            result = await session.execute(query)
-            return result.scalar_one_or_none()
+            row = await session.execute(query)
+            row = row.scalar_one_or_none()
+            if row is None:
+                raise TheReaderWasNotFoundException
+            return row
